+++ src/navigation/types.ts
@@ -146,6 +146,8 @@ export type TrueSheetNavigationSheetProps = Pick<
   | 'anchor'
   | 'anchorOffset'
   | 'elevation'
+  | 'detached'
+  | 'detachedOffset'
   | 'stackBehavior'
 >;